user_management/requirements.txt
# Django web framework
Django==4.2.3
psycopg2-binary
python-dotenv
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
django-staticfiles==1.2.1
Pillow
requests
